chuyển beHurt(int) -> beHurt(float)

+++ src/main/java/com/cyberpunk/Object/MapObject.java
@@ -323,7 +323,7 @@ public class MapObject extends Object {
 	}
 
 	@Override
-	public void beHurt(int damageGet) {
+	public void beHurt(float damageGet) {
 		if (getState() == NOBEHURT) {
 			return;
 		}
